Removing centering behaviour in responsive mode
Hi there

Thanks for adding in the class on the div to allow centering. Yes works
fine (although not the CDN version - needs to be updated too?). However
personally I would prefer just to have the class included but not the
CSS. Sometimes I want the videos centered, sometimes I don't, so to me
it just seems cleaner to leave out the CSS and allow people to add it
via a child theme stylesheet if they want to. Up to you though it can
work either way I guess...

Cheers
Jon